reverting back to non-oidc based authentication
