+++ config/auth/auth0.config.js
@@ -11,6 +11,6 @@ import {
 export const AUTH0_CLIENT_ID :string =
   isProd
     ? 'o8Y2U2zb5Iwo01jdxMN1W2aiN8PxwVjh'
-    : 'uy3biAScODcrQnSZIVLD4Auc39K4ioB1';
+    : 'KTzgyxs6KBcJHB872eSMe2cpTHzhxS99';
 
 export const AUTH0_DOMAIN :string = 'openlattice.auth0.com';
