+++ src/index.js
@@ -39,6 +39,14 @@ injectGlobal`
 `;
 /* eslint-enable */
 
+/*
+ * // !!! MUST HAPPEN FIRST !!!
+ */
+Auth0.initialize();
+/*
+ * // !!! MUST HAPPEN FIRST !!!
+ */
+
 const routerHistory = initializeRouterHistory();
 const reduxStore = initializeReduxStore(routerHistory);
